Self-play: ship action descriptors narrow (i64->i8/i16), +10.9%
The per-action descriptors held tiny values shipped as i64: a_type ∈ {-1,0,1,2}, a_pidx ∈
[0,675], a_ltok ∈ [0,7]. Narrow them to i8/i16/i8 in the netbatch collect() (5 bytes/entry vs
40) and cast back to i64 on the GPU for torch.gather. Bit-exact (values fit; verified gather
output finite, no NaN). Also shrinks the numpy arrays the collect marshals.

Measured (RTX 4070S, 128 sims, batch 2048): 205,355 -> 227,742 leaf-evals/s (+10.9%), better
than the ~6% the profile implied. Cumulative vs baseline: 95,070 -> 227,742 = 2.40x.

+++ agent/engine-bridge/src/batch_selfplay.rs
@@ -633,18 +633,20 @@ impl BatchedNetSelfPlay {
 
         // Snapshot leaf states + build per-action descriptors (cheap, sequential).
         let mut states: Vec<GameState> = Vec::with_capacity(b);
-        let mut a_type = vec![-1i64; b * amax];
-        let mut a_pidx = vec![0i64; b * amax];
-        let mut a_ltok = vec![0i64; b * amax];
+        // Descriptors ship narrow (5 bytes/entry vs 40 as i64) and are cast back to i64 on the
+        // GPU for `gather`: a_type ∈ {-1,0,1,2}, a_pidx ∈ [0,675], a_ltok ∈ [0,7].
+        let mut a_type = vec![-1i8; b * amax];
+        let mut a_pidx = vec![0i16; b * amax];
+        let mut a_ltok = vec![0i8; b * amax];
         let mut a_mask = vec![0u8; b * amax];
         for (i, &slot) in self.pending.iter().enumerate() {
             let node = &self.games[slot].arena[self.games[slot].pending_leaf as usize];
             states.push(node.gs);
             for (j, &act) in node.actions.iter().enumerate() {
                 let (t, p, l) = action_descriptor(act, node.gs.phase);
-                a_type[i * amax + j] = t;
-                a_pidx[i * amax + j] = p;
-                a_ltok[i * amax + j] = l;
+                a_type[i * amax + j] = t as i8;
+                a_pidx[i * amax + j] = p as i16;
+                a_ltok[i * amax + j] = l as i8;
                 a_mask[i * amax + j] = 1;
             }
         }
